requirements.txt
streamlit
Pillow
transformers
torch
torchvision
requests
snowflake-connector-python
pandas
plotly
imagehash
numpy
scikit-image
uagents
pyarrow<19.0.0
